requirements.txt
tensorflow
pytorch
keras
